+++ odl/space/__init__.py
@@ -25,13 +25,9 @@ from . import base_ntuples
 
 from . import weighting
 
-from . import ntuples
-from .ntuples import *
-__all__ += ntuples.__all__
-
-from . import cu_ntuples
-from .cu_ntuples import *
-__all__ += cu_ntuples.__all__
+from . import npy_ntuples
+from .npy_ntuples import *
+__all__ += npy_ntuples.__all__
 
 from . import pspace
 from .pspace import *
@@ -40,6 +36,10 @@ __all__ += pspace.__all__
 from . import fspace
 from .fspace import *
 __all__ += fspace.__all__
+
+from . import entry_points
+from .entry_points import *
+__all__ += entry_points.__all__
 
 from . import space_utils
 from .space_utils import *